Added immunity penalization for the event scheduler.
git-svn-id: https://openmalaria.googlecode.com/svn/trunk@593 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ model/WithinHost/WithinHostModel.cpp
@@ -176,8 +176,8 @@ void WithinHostModel::updateImmuneStatus(){
 }
 
 void WithinHostModel::immunityPenalisation() {
-  _cumulativeY=(double)_cumulativeYlag-(immPenalty_22*(_cumulativeY-_cumulativeYlag));
-  if (_cumulativeY <  0) {
+  _cumulativeY = _cumulativeYlag - immPenalty_22*(_cumulativeY-_cumulativeYlag);
+  if (_cumulativeY < 0) {
     _cumulativeY=0.0;
   }
 }
